+++ tests/api_send/test_send_transfer.py
@@ -14,10 +14,10 @@
 # limitations under the License.
 
 from tests.api_send.test_send_super import TestSendSuper
-from IconService.builder.transaction_builder import TransactionBuilder
-from IconService.signed_transaction import SignedTransaction
-from IconService.utils.validation import is_icx_transaction, is_T_HASH
-from IconService.exception import JSONRPCException, DataTypeException
+from iconsdk.builder.transaction_builder import TransactionBuilder
+from iconsdk.signed_transaction import SignedTransaction
+from iconsdk.utils.validation import is_icx_transaction, is_T_HASH
+from iconsdk.exception import JSONRPCException, DataTypeException
 
 
 class TestSendTransfer(TestSendSuper):