+++ tests/api_send/test_signed_transaction.py
@@ -14,10 +14,10 @@
 # limitations under the License.
 
 from unittest import main
-from IconService.signed_transaction import SignedTransaction
-from IconService.builder.transaction_builder import TransactionBuilder, MessageTransactionBuilder, \
+from iconsdk.signed_transaction import SignedTransaction
+from iconsdk.builder.transaction_builder import TransactionBuilder, MessageTransactionBuilder, \
     CallTransactionBuilder, DeployTransactionBuilder
-from IconService.utils.validation import is_icx_transaction, is_call_transaction, is_message_transaction, \
+from iconsdk.utils.validation import is_icx_transaction, is_call_transaction, is_message_transaction, \
     is_deploy_transaction, is_T_HASH
 from tests.api_send.test_send_super import TestSendSuper