+++ tests/example_config.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-#TEST_HTTP_ENDPOINT_URI_V3 = "http://52.79.233.89:9000/api/v3"
-TEST_HTTP_ENDPOINT_URI_V3 = "http://localhost:9000/api/v3"
+TEST_HTTP_ENDPOINT_URI_V3 = "http://13.209.103.183:9000/api/v3"
+#TEST_HTTP_ENDPOINT_URI_V3 = "http://localhost:9000/api/v3"
 
 TEST_PRIVATE_KEY = bytes.fromhex("2d42994b2f7735bbc93a3e64381864d06747e574aa94655c516f9ad0a74eed79")