+++ tests/libs/test_serializer.py
@@ -15,7 +15,7 @@
 
 from hashlib import sha3_256
 from unittest import TestCase, main
-from IconService.libs.serializer import serialize
+from iconsdk.libs.serializer import serialize
 from tests.example_tx_requests import TEST_REQUEST_TRANSFER_ICX, TEST_REQUEST_SCORE_FUNCTION_CALL