+++ tests/utils/test_validation.py
@@ -15,9 +15,9 @@
 
 from os import path
 from unittest import TestCase, main
-from IconService.utils.validation import is_keystore_file, has_keys, is_keystore_file_for_icon
+from iconsdk.utils.validation import is_keystore_file, has_keys, is_keystore_file_for_icon
 from eth_keyfile import load_keyfile
-from IconService.exception import KeyStoreException
+from iconsdk.exception import KeyStoreException
 
 
 class TestValidation(TestCase):