+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+
+## [1.0.1] -  August, 30, 2018
+### Added
+N/A
+
+### Changed
+- Revise the package names and update the version to 1.0.1
+    - `IconService` to `iconsdk`
+    - `Icon_service` to `icon_service`
+
+### Removed
+N/A
+
+
 ## [1.0.0] -  July, 31, 2018
 ### Added
  - Initialize version.
@@ -9,3 +23,7 @@ N/A
 
 ### Removed
 N/A
+
+
+
+